Enabling to remove all timers + Adding a global channel timer displaying the time left since the most reacent spawn +updating the horse counter to have bigger text size and switching from a text box to numerical counter

+++ HorseTrack/Models/ChannelInformation.cs
@@ -4,6 +4,6 @@
     {
         public string ChannelName { get; set; }
 
-        public int[] Times { get; set; }
+        public long[] Times { get; set; }
     }
 }
